Ticket 132 - reduce use of DHCPv6
Also fixed a few minor textual issues noticed while scanning for DHCPv6 usage.

Note: DHCPv6 usage early in the document remains as the terminology has not yet been defined and it likely reads better (this is closer to what was done in RFC 3315).

Please use "DHCP" instead of "DHCPv6" when updating all but the early sections of the document (obviously use DHCPv6 when providing the title of documents or where needed for clarity).
# Conflicts:
#	rfc3315bis.xml