Cleanup Option vs option usage, and a bit more
General	"There are places where we use ""IA_PD Prefix option"" vs ""IA_PD option"" or perhaps the intent was to reference the ""IA Prefix option"" (i.e., IAPREFIX). We need to clean this up to either use IA_PD option or IA Prefix option depending on which was intended.

Also, corrected IA Address and IA Prefix options to be clear that each only holds one address or prefix (respectively). This is in first paragraph of the option definitions sections."