Address more of Jinmei's comments
Section 17.1.10.1, "This bullet:
   -  Sends a Request message if any of the IAs in the Reply message
      contains the NoBinding status code.  [...]

and this sentence seem to contradict each other.

   [...]  This facilitates the client using a single state machine
   for all bindings.

  Specifically, the bullet description seems to lead to multiple separate state machines for different bindings." Jinmei, Tatuya (https://www.ietf.org/mail-archive/web/dhcwg/current/msg17577.html)

Section 17.2.2, "[...]  If the reconfigure mechanism is supported, the server is supposed to send Authentication option with Reconfigure Key (see Section 19.4 for details).

What if we re-introduce securer authentication (whether it's sedhcpv6 or not)?  Should we still keep using Reconfigure Key, which contains a private key in plain text?  (see also comment on Section 19.4)"	Jinmei, Tatuya (https://www.ietf.org/mail-archive/web/dhcwg/current/msg17577.html)

Section 17.2.9, "  [...]  If the Option Request option includes a container option the server MUST include all the options that are eligible to be encapsulated in the container.

  Is the term ""container option"" defined somewhere?  If not maybe it should in the terminology section.", Jinmei, Tatuya (https://www.ietf.org/mail-archive/web/dhcwg/current/msg17577.html)