Address more of Tomek's comments
General Comment	- "The header appearing on each page is ""RFC 3315 bis"". It's ok for the time this draft is discussed in WG phase, but we should change this to ""DHCP for IPv6"" before sending it to the IESG.", Tomek Mrugalski https://www.ietf.org/mail-archive/web/dhcwg/current/msg17619.html

Section 21:
I don't understand this paragraph: "In the case where relay agents add additional options to Relay Forward messages, the messages exchanged between relay agents and servers may be used to mount a "man in the middle or denial of service attack."

It doesn't make sense to me. If a relay agent is compromised (or attackers manages to introduce a node he controls into the path, both DOS and MITM attacks are possible, regardless of the original relay agent inserted any options or not.

"in which the DHCP server sends the key to the client." => "in which the DHCP server sends the key in plain text to the client."

"A malicious requesting router may be able to mount a denial of service attack by repeated requests for delegated prefixes that exhaust the delegating router's available prefixes." We should add something similar to: "Some forms of this exhaustion attack can be mitigated by appropriate server policy, e.g. limiting maximum number of leases one client can get".

Unless I missed something, we can probably remove this sentence "The details of using IPsec for DHCP are under development."". Or replace it with a link to draft-ietf-dhc-sedhcpv6.", Tomek Mrugalski https://www.ietf.org/mail-archive/web/dhcwg/current/msg17619.html