More on Ticket #167 based on comments from Lorenzo
Lorenzo wanted fewer references to requesting router and delegating router. So, reworked section 6.3 to move last paragraph about using client/server instead earlier and replace requesting router/delegating router in remainder of section.

However, not sure if it now reads as well as it gets a bit messy when talking about DHCP server in the client.