Removed server-side text from 18.2.4 (Renew)
17.1.4	"""The server determines new
   lifetimes for the leases according to the administrative
   configuration of the server.  The server may also add leases to the
   IAs.  The server can remove leases from the IAs by returning IA
   Address options (for IA_NA and IA_TA) and IA Prefix options (for
   IA_PD) with preferred and valid lifetimes set to 0.""

This text is server specific and should be better moved to the server
section, or alternatively we may want to consider updating this text to
be more of a client-side type, e.g.

""The client MUST be prepared that the server removes or adds new leases
to the IAs. The leases are removed when server sets their preferred and
valid lifetimes to 0, in which case the client MUST stop using them immediately.""
"	Marcin Siodelski, https://www.ietf.org/mail-archive/web/dhcwg/current/msg17576.html	10/19 - Tomek, Marcin, and Bernie agreed that this was a good change to make 17.1 text client specific; perhaps move server specific actions into 17.2 as appropriate.	"Bernie - Done.

I decided to REMOVE this text. I don't really see any benefit of this text (either what was there or the proposed client-side text). The text in 18.2.10.1, Reply for Solicit (with Rapid Commit), Request, Renew or Rebind, is pretty clear about all of these things that can happen so it seems redundant (less is better)."