Minor edits
Changed in section 13.2, "So, it is NOT RECOMMENDED for a client to renew expired temporary" to remove expired, since that is a bit ambiguous. A client can renew at any time (might not be expired) and does expired refer to preferred lifetime or valid lifetime (since if address is no longer valid, not really "renewed").

Changed "incidate" to "indicate" just before 18.3.

And a few others noticed while comparing against the 08.