IETF LC, incorporate RFC7283
Minor Issues, Elwyn Davies, https://www.ietf.org/mail-archive/web/ietf/current/msg106146.html

Non-incorporation of RFC 7283:  Is there a good reason why the update of relay agent behaviour (which is apparently of general application across DHCPv6) is not incorporated into this document as with several other items?  The update is short and would be readily addressed in s19 here.  As it is one is left with another document with some key modifications left around rather than achieving the unification aim of this -bis