Start work on adding references for options consistently
Nits, Elwyn Davies, https://www.ietf.org/mail-archive/web/ietf/current/msg106146.html

s18 (multiple places):  In many places the first reference to an option in a subsection is accompanied by a pointer to the part of s21 that defines it (Good!).  However this is not consistent.  I have picked up several cases where the reference is missing but there may be others.  Please check and make it consistent.  However, in the case of Client Identifier and Server Identifier options and the IA options, they appear so many times that it might be worth just adding a paragraph about the identifiers with the reference pointers in Section 18.	

Note: We will add a "(see Section 21.x)" reference to the FIRST usage of an option in any section - across the entire document.